refactor(mediawiki): extract content validation from errors.go (deslop #59)
